grouped semaphores into sets. fixed send/receive signals.

+++ MiUtil/MiScreen.cs
@@ -18,13 +18,13 @@ namespace MiUtil
             Visible = false;
         }
 
-        public virtual IEnumerator<int> EntrySequence() { yield return 0; }
+        public virtual IEnumerator<ulong> EntrySequence() { yield return 0; }
 
-        public virtual IEnumerator<int> Upped() { yield return 0; }
-        public virtual IEnumerator<int> Downed() { yield return 0; }
-        public virtual IEnumerator<int> Lefted() { yield return 0; }
-        public virtual IEnumerator<int> Righted() { yield return 0; }
-        public virtual IEnumerator<int> Pressed() { yield return 0; }
-        public virtual IEnumerator<int> Cancelled() { yield return 0; }
+        public virtual IEnumerator<ulong> Upped() { yield return 0; }
+        public virtual IEnumerator<ulong> Downed() { yield return 0; }
+        public virtual IEnumerator<ulong> Lefted() { yield return 0; }
+        public virtual IEnumerator<ulong> Righted() { yield return 0; }
+        public virtual IEnumerator<ulong> Pressed() { yield return 0; }
+        public virtual IEnumerator<ulong> Cancelled() { yield return 0; }
     }
 }
